src/docker/django/requirements.txt
-i https://pypi.org/simple
asgiref==3.5.2
django==4.1.4
sqlparse==0.4.3
click==8.1.3
django-click==2.3.0
django-pytest